requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-Login==0.5.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
flask-restx==0.2.0
flask-accepts==0.16.6

pytest
marshmallow
Flask-Script
psycopg2-binary